attendee, attendees, events and newEventsForms moved from app.js to their components
